+++ src/oop/Main.java
@@ -3,7 +3,7 @@ package oop;
 public class Main {
     public static void main(String[] args) {
         CerejaBicicletas bikeBase = new CerejaBicicletas();
-        bikeBase.setMarcha(7);
+        bikeBase.setMarcha((byte) 7);
         bikeBase.getMarcha();
         bikeBase.getVelocidade();
         bikeBase.aumentaAVelocidade(8);
@@ -11,6 +11,6 @@ public class Main {
         bikeBase.aumentaAVelocidade(8);
         bikeBase.getVelocidade();
         BikeDeMontanha bikeDeMontanha = new BikeDeMontanha();
-        bikeDeMontanha.setMarcha(8);
+        bikeDeMontanha.setMarcha((byte) 8);
     }
 }
